t/001-init.t: less restrictive timezone check
The DOCHAZKA_TIMEZONE config param must be set to a string that
the PostgreSQL server recognizes as a timezone.

Since we have no easy way of knowing what strings the server
will so recognize, and now that our tests are no longer hard-coded
for Central European Time, it is enough to check that DOCHAZKA_TIMEZONE is
set to *something* (i.e. not undef or empty).

Signed-off-by: Nathan Cutler <ncutler@suse.com>