Test.pm: initialize_unit() return $app instead of $status
The first thing initialize_unit() does is to call Web::MREST::init(),
which returns a status object. If this is not OK, the entire unit is
considered compromised and "plan skipall" is called with the text of the
error message. When the status is OK, the payload contains the name of the
application. Return this string instead of the entire status object.

Signed-off-by: Nathan Cutler <ncutler@suse.com>